refactor: ImageSizes and move vue types to types/
