+++ src/runtime/components/nuxt-picture.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script lang="ts">
-import type { DefineComponentWithMixin } from './types'
+import type { DefineComponentWithMixin } from '../../types/vue'
 
 import { imageMixin } from './image.mixin'
 import { getFileExtension } from '~image'
